Location changed
Tested locally the wrapping interactor Housings::FetchDetailsAndUpdate, no issues
