+++ lib/housing_list/interactors/housings/fetch_details_and_update.rb
@@ -15,11 +15,11 @@ module Housings
 
     def call
       provider_name       = housing.provider.split('_').map(&:capitalize).join
-      scraper_class_name  = "Housings::Scrapers::#{provider_name}"
+      scraper_class_name  = "Scrapers::#{provider_name}"
       scraper_class       = Object.const_get(scraper_class_name)
-      result              = scraper_class.new(url: housing.url).call
+      housing_attributes  = scraper_class.new(url: housing.url).scrape
 
-      HousingRepository.new.update(housing.id, result.housing_attributes.merge(imported: true))
+      HousingRepository.new.update(housing.id, housing_attributes.merge(imported: true))
     end
   end
 end
